tiny changes

+++ Periode1/Predictif-Web/src/main/java/fr/insalyon/predictif/ihm/web/action/DemanderConsultationsAction.java
@@ -27,7 +27,7 @@ public class DemanderConsultationsAction extends Action {
         String mediumName = request.getParameter("medium");
         if(id!=null) {
             //System.out.println("client");
-            c = service.rechercherClientParId((long) id);
+            c = service.rechercherClientParId(id);
             if(!service.inConsultation(c)) {
                 if(mediumName!=null) {
                     m = service.chercherMedium(mediumName);
